:sparkles: Déplacement de getRuleFromAnalysis dans le moteur
