+++ source/selectors/analyseSelectors.js
@@ -314,19 +314,3 @@ export let currentQuestionSelector = createSelector(
 	(nextSteps, unfoldedStep) => unfoldedStep || head(nextSteps)
 )
 
-export let getRuleFromAnalysis = analysis => dottedName => {
-	if (!analysis) {
-		throw new Error("[getRuleFromAnalysis] The analysis can't be nil !")
-	}
-	let rule =
-		analysis.cache[dottedName]?.explanation || // the cache stores a reference to a variable, the variable is contained in the 'explanation' attribute
-		analysis.targets.find(propEq('dottedName', dottedName))
-
-	if (!rule) {
-		throw new Error(
-			`[getRuleFromAnalysis] Unable to find the rule ${dottedName}`
-		)
-	}
-
-	return rule
-}
